+++ crates/cubed-application/src/use_cases/mod.rs
@@ -1,6 +1,7 @@
 pub mod create_server;
 pub mod delete_server;
 pub mod init_file_system;
+pub mod reserve_port;
 pub mod restart_server;
 pub mod select_java;
 pub mod start_server;
@@ -9,6 +10,7 @@ pub mod stop_server;
 pub use create_server::{CreateServer, CreateServerInput};
 pub use delete_server::DeleteServer;
 pub use init_file_system::InitFileSystem;
+pub use reserve_port::ReservePort;
 pub use restart_server::RestartServer;
 pub use select_java::SelectJava;
 pub use start_server::StartServer;